Begin a new history for your answers
Unlike the branches you've seen so far in class, this branch points to a
history that's completely separate from the one all the other branches
reference--each group of commits is unreachable from the other.

This is a perfectly valid setup, although in most cases you'll probably
want to have multiple repositories instead of multiple histories in one
repository.